Fix CodeRabbit issues: separate classes, fix package names, update to jakarta validation, implement constructor injection
