+++ backend/src/main/java/com/marocotransport/controller/TruckController.java
@@ -1,4 +1,4 @@
-package com.maroctransit.auth.controller;
+package com.marocotransport.controller;
 
 import com.maroctransit.auth.dto.ApiResponse;
 import com.maroctransit.auth.dto.TruckDTO;
@@ -8,13 +8,13 @@ import com.maroctransit.auth.service.TruckService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.util.List;
 
 /**
@@ -25,8 +25,11 @@ import java.util.List;
 @Tag(name = "Truck Management", description = "APIs for managing trucks and their availability")
 public class TruckController {
 
-    @Autowired
-    private TruckService truckService;
+    private final TruckService truckService;
+
+    public TruckController(TruckService truckService) {
+        this.truckService = truckService;
+    }
 
     /**
      * Register a new truck (Carrier only)
@@ -147,105 +150,4 @@ public class TruckController {
                     .body(ApiResponse.error("Truck not found or you don't have permission to delete"));
         }
     }
-}npackage com.maroctransit.auth.controller;
-
-import com.maroctransit.auth.dto.ApiResponse;
-import com.maroctransit.auth.dto.MatchingRequestDTO;
-import com.maroctransit.auth.dto.MatchingResultDTO;
-import com.maroctransit.auth.service.MatchingService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.List;
-
-/**
- * REST Controller for ML-based job-truck matching
- */
-@RestController
-@RequestMapping("/api/v1/matching")
-@Tag(name = "ML Matching", description = "APIs for ML-based carrier-shipper matching")
-public class MatchingController {
-
-    @Autowired
-    private MatchingService matchingService;
-
-    /**
-     * Match jobs to available trucks using ML
-     */
-    @PostMapping("/jobs-to-trucks")
-    @PreAuthorize("hasRole('SHIPPER')")
-    @Operation(
-        summary = "Match jobs to trucks", 
-        description = "Find best trucks for specific jobs using ML matching algorithm",
-        security = @SecurityRequirement(name = "bearerAuth")
-    )
-    public ResponseEntity<ApiResponse<List<MatchingResultDTO>>> matchJobsToTrucks(
-            @Valid @RequestBody MatchingRequestDTO matchingRequest) {
-        
-        List<MatchingResultDTO> matchingResults = matchingService.matchJobsToTrucks(
-                matchingRequest.getJobIds(), 
-                matchingRequest.getParameters());
-        
-        return ResponseEntity.ok(ApiResponse.success("Matching completed successfully", matchingResults));
-    }
-
-    /**
-     * Match trucks to available jobs using ML
-     */
-    @PostMapping("/trucks-to-jobs")
-    @PreAuthorize("hasRole('CARRIER')")
-    @Operation(
-        summary = "Match trucks to jobs", 
-        description = "Find best jobs for specific trucks using ML matching algorithm",
-        security = @SecurityRequirement(name = "bearerAuth")
-    )
-    public ResponseEntity<ApiResponse<List<MatchingResultDTO>>> matchTrucksToJobs(
-            @Valid @RequestBody MatchingRequestDTO matchingRequest) {
-        
-        List<MatchingResultDTO> matchingResults = matchingService.matchTrucksToJobs(
-                matchingRequest.getTruckIds(), 
-                matchingRequest.getParameters());
-        
-        return ResponseEntity.ok(ApiResponse.success("Matching completed successfully", matchingResults));
-    }
-
-    /**
-     * Get matching recommendations based on historical data
-     */
-    @GetMapping("/recommendations")
-    @PreAuthorize("hasAnyRole('SHIPPER', 'CARRIER')")
-    @Operation(
-        summary = "Get recommendations", 
-        description = "Get personalized recommendations based on historical data",
-        security = @SecurityRequirement(name = "bearerAuth")
-    )
-    public ResponseEntity<ApiResponse<List<MatchingResultDTO>>> getRecommendations(
-            @RequestParam(defaultValue = "5") int limit) {
-        
-        List<MatchingResultDTO> recommendations = matchingService.getRecommendations(limit);
-        
-        return ResponseEntity.ok(ApiResponse.success("Recommendations generated successfully", recommendations));
-    }
-
-    /**
-     * Get matching analytics and insights
-     */
-    @GetMapping("/analytics")
-    @PreAuthorize("hasAnyRole('SHIPPER', 'CARRIER', 'ADMIN')")
-    @Operation(
-        summary = "Get matching analytics", 
-        description = "Get analytics and insights about matching patterns",
-        security = @SecurityRequirement(name = "bearerAuth")
-    )
-    public ResponseEntity<ApiResponse<Object>> getMatchingAnalytics() {
-        Object analytics = matchingService.getMatchingAnalytics();
-        
-        return ResponseEntity.ok(ApiResponse.success("Analytics generated successfully", analytics));
-    }
-}MatchingController.java
+}
