+++ backend/src/main/java/com/marocotransport/security/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.maroctransit.auth.security;
+package com.marocotransport.security;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
